% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn2-data.R
\docType{data}
\name{bn2_data}
\alias{bn2_data}
\title{Between Subjects Twoway ANOVA Example Data}
\format{A data frame of ratings of close interpersonal attachments

year: fiscal year - either 2007 or 2008
type: varying sports - baseball, basketball, football, soccer, volleyball
money: athletic spending (in dollars)}
\usage{
data(bn1_data)
}
\description{
Dataset for use in \code{\link{omega.partial.SS.bn}},
including two year's worth of athletic spending data
(treating each receipt and years as separate
between subjects' events) for five different sports.
}
\references{
Nolan and Heizen Statistics for the Behavioral Sciences
(\href{https://www.macmillanlearning.com/Catalog/product/statisticsforthebehavioralsciences-rentalonly-fourthedition-nolan}{Book Link})
}
\keyword{datasets}
