% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dept.avg.R
\name{d.dept.avg}
\alias{d.dept.avg}
\title{d.dept.avg}
\usage{
d.dept.avg(m1, m2, sd1, sd2, n, a = 0.05, k = 2)
}
\arguments{
\item{m1}{mean from first group}

\item{m2}{mean from second group}

\item{sd1}{standard deviation from first group}

\item{sd2}{standard deviation from second group}

\item{n}{sample size}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function displays cohen's d and confidence interval for dependent t from averages
}
\examples{
d.dept.avg(m1=20, m2=17, sd1=4, sd2=5, n=100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
