% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dept.diff.R
\name{d.dept.diff}
\alias{d.dept.diff}
\title{d.dept.diff}
\usage{
d.dept.diff(mdiff, sddiff, n, a = 0.05, k = 2)
}
\arguments{
\item{mdiff}{mean difference score}

\item{sddiff}{standard deviation of the differences scores}

\item{n}{sample size}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function displays Cohen's d and confidence itnerval for dependent t differences
}
\examples{
d.dept.diff(mdiff = 5, sddiff = 1, n=100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
