% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.ind.t.t.R
\name{d.ind.t.t}
\alias{d.ind.t.t}
\title{d.ind.t.t}
\usage{
d.ind.t.t(t, n1, n2, a = 0.05)
}
\arguments{
\item{t}{t-test value}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\description{
This function displays d for between subjects data
and the non-central confidence interval estimating from the t-statistic.
}
\examples{
d.ind.t.t(t = 4.12, n1 = 100, n2 = 100, a = .05)
}
\keyword{effect}
\keyword{independent}
\keyword{size,}
\keyword{t}
