% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.ind.tt.R
\name{d.ind.tt}
\alias{d.ind.tt}
\title{d.ind.tt}
\usage{
d.ind.tt(t, n1, n2, a = 0.05, k = 2)
}
\arguments{
\item{t}{t-test value}

\item{n1}{sample size from first group}

\item{n2}{sample size from second group}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays Cohen's d and confidence interval for independent t from t.
}
\examples{
d.ind.tt(t = 4.12, n1 = 100, n2 = 100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
