% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.single.t.R
\name{d.single.t}
\alias{d.single.t}
\title{d.single.t}
\usage{
d.single.t(m, u, sd, n, a = 0.05, k = 2)
}
\arguments{
\item{m}{sample mean}

\item{u}{population mean}

\item{sd}{sample standard deviation}

\item{n}{sample size}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays Cohen's d and confidence interval for single t from means.
}
\examples{
d.single.t(m=20, u=17, sd=4, n=100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
