% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.single.t.t.R
\name{d.single.t.t}
\alias{d.single.t.t}
\title{d for Single t from t}
\usage{
d.single.t.t(t, n, a = 0.05)
}
\arguments{
\item{t}{t-test value}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d and non-central confidence interval for single t
estimated from the t-statistic.
}
\examples{
d.single.t.t(t = 4.20, n = 100, a = .05)
}
\keyword{effect}
\keyword{single}
\keyword{size,}
\keyword{t}
