% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.single.tt.R
\name{d.single.tt}
\alias{d.single.tt}
\title{d.single.tt}
\usage{
d.single.tt(t, n, a = 0.05, k = 2)
}
\arguments{
\item{t}{t-test value}

\item{n}{sample size}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays Cohen's d and confidence interval for single t from t.
}
\examples{
d.single.tt(t=4.20, n=100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
