% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.to.r.R
\name{d.to.r}
\alias{d.to.r}
\title{r and Coefficient of Determination (R2) from d}
\usage{
d.to.r(d, n1, n2, a = 0.05)
}
\arguments{
\item{d}{effect size statistic}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\description{
Calculates r from d and then translates r to r2 to calculate
the non-central confidence interval for r2 using the F distribution.
}
\examples{
d.to.r(d = .5, n1 = 50, n2 = 50, a = .05)
}
\keyword{correlation}
\keyword{effect}
\keyword{size,}
