% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.z.mean.R
\name{d.z.mean}
\alias{d.z.mean}
\title{d.z.mean}
\usage{
d.z.mean(mu, m1, sig, sd1, n, a = 0.05)
}
\arguments{
\item{mu}{population mean}

\item{m1}{sample study mean}

\item{sig}{population standard deviation}

\item{sd1}{standard deviation from the study}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for z-scores with the 
population mean and standard deviation.
The normal confidence interval is also provided.
}
\examples{
d.z.mean(mu = 20, m1 = 17, sig = 4, sd1 = 5, n = 100, a = .05)
}
\keyword{effect}
\keyword{size,}
\keyword{z-test}
