% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.ind.t.R
\name{delta.ind.t}
\alias{delta.ind.t}
\title{delta.ind.t}
\usage{
delta.ind.t(m1, m2, sd1, sd2, n1, n2, a = 0.05)
}
\arguments{
\item{m1}{mean from control group}

\item{m2}{mean from experimental group}

\item{sd1}{standard deviation from control group}

\item{sd2}{standard deviation from experimental group}

\item{n1}{sample size from control group}

\item{n2}{sample size from experimental group}

\item{a}{significance level}
}
\description{
This function displays d-delta for between subjects data
and the non-central confidence interval using the
control group standard deviation as the denominator.
}
\examples{
delta.ind.t(m1 = 20, m2 = 17, sd1 = 4, sd2 = 5, n1 = 100, n2 = 100, a = .05)
}
\keyword{delta,}
\keyword{effect}
\keyword{independent}
\keyword{size,}
\keyword{t}
