% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dept-data.R
\docType{data}
\name{dept_data}
\alias{dept_data}
\title{Dependent t Example Data}
\format{
A data frame of before and after scores for rating
supernatural beliefs.

before: scores rated before watching a scifi movie
after: scores rated after watching a scifi movie
}
\usage{
data(dept_data)
}
\description{
Dataset for use in \code{\link{d_dep_t_diff}}, \code{\link{d_dep_t_diff_t}},
\code{\link{d_dep_t_avg}}, and \code{\link{d_dep_t_rm}} exploring the before
and after effects of scifi movies on supernatural beliefs.
}
\references{
Nolan and Heizen Statistics for the Behavioral Sciences
\href{https://www.macmillanlearning.com/Catalog/product/statisticsforthebehavioralsciences-rentalonly-fourthedition-nolan}{Book Link}
}
\keyword{datasets}
