% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Eta.anova.R
\name{Eta.anova}
\alias{Eta.anova}
\title{Eta.anova}
\usage{
Eta.anova(dfm, dfe, f, a = 0.05, k = 2)
}
\arguments{
\item{dfm}{degrees of freedom model}

\item{dfe}{degrees of freedom error}

\item{f}{F statistic}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays eta, r squared, ICCs from ANOVA analyses
}
\examples{
Eta.anova(dfm = 2, dfe = 20, f = 5.7, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
