% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.ind.t.R
\name{g.ind.t}
\alias{g.ind.t}
\title{g.ind.t}
\usage{
g.ind.t(m1, m2, sd1, sd2, n1, n2, a = 0.05, k = 2)
}
\arguments{
\item{m1}{mean from first group}

\item{m2}{mean from second group}

\item{sd1}{standard deviation from first group}

\item{sd2}{standard deviation from second group}

\item{n1}{sample size from first group}

\item{n2}{sample size from second group}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays Hedge's g and confidence interval for independent t from means.
}
\examples{
g.ind.t(m1=20, m2=17, sd1=4, sd2=5, n1=100, n2=100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
