% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.partial.SS.mix.R
\name{ges.partial.SS.mix}
\alias{ges.partial.SS.mix}
\title{Partial Generalized Eta-Squared for Mixed Design ANOVA from F}
\usage{
ges.partial.SS.mix(dfm, dfe, ssm, sss, sse, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{ssm}{sum of squares for the model/IV/between}

\item{sss}{sum of squares subject variance}

\item{sse}{sum of squares for the error/residual/within}

\item{Fvalue}{F statistic}

\item{a}{significance level}
}
\value{
Partial generalized eta-squared (GES) with associated confidence intervals
and relevant statistics.
\item{ges}{effect size}
\item{geslow}{lower level confidence interval for ges}
\item{geshigh}{upper level confidence interval for ges}
\item{dfm}{degrees of freedom for the model/IV/between}
\item{dfe}{degrees of freedom for the error/residual/within}
\item{F}{F-statistic}
\item{p}{p-value}
}
\description{
This function displays partial generalized eta-squared (GES) from ANOVA analyses
and its non-central confidence interval based on the F distribution.
This formula works for mixed designs.
}
\examples{
ges.partial.SS.mix(dfm = 2, dfe = 100, ssm = 435, sss = 235, sse = 659, Fvalue = 5.46, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{ges,}
\keyword{size,}
