% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.partial.SS.rm.R
\name{ges.partial.SS.rm}
\alias{ges.partial.SS.rm}
\title{ges.partial.SS.rm}
\usage{
ges.partial.SS.rm(dfm, dfe, ssm, sss, sse1, sse2, sse3, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{ssm}{sum of squares for the model/IV/between}

\item{sss}{sum of squares subject variance}

\item{sse1}{sum of squares for the error/residual/within for the first IV}

\item{sse2}{sum of squares for the error/residual/within for the second IV}

\item{sse3}{sum of squares for the error/residual/within for the interaction}

\item{Fvalue}{F statistic}

\item{a}{significance level}
}
\description{
This function displays partial ges squared from ANOVA analyses
and its non-central confidence interval based on the F distribution.
This formula works for multi-way repeated measures designs.
}
\examples{
ges.partial.SS.rm(dfm = 2, dfe = 100, ssm = 435, sss = 659, sse1 = 435, sse2 = 446, sse3 = 546, Fvalue = 5.46, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{ges,}
\keyword{size,}
