% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NcpChisq.R
\name{NcpChisq}
\alias{NcpChisq}
\title{NcpChisq}
\usage{
NcpChisq(x, q, df, confirm = FALSE)
}
\arguments{
\item{x}{chisquare value}

\item{q}{input}

\item{df}{degrees of freedom}

\item{confirm}{defaults to FALSE}
}
\description{
This Function calculates non-central chi-square
}
\examples{
NcpChisq(x, q, df, confirm=FALSE)
}
\keyword{effect}
\keyword{size}
