% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncpf.R
\name{Ncpf}
\alias{Ncpf}
\title{Ncpf}
\usage{
Ncpf(x, q, df1, df2, confirm = FALSE)
}
\arguments{
\item{x}{value}

\item{q}{value}

\item{df1}{degrees of freedom model}

\item{df2}{degrees of freedom error}

\item{confirm}{defaults to FALSE}
}
\description{
This Function calculates non-central F
}
\examples{
Ncpf(x, q, df1, df2, confirm=FALSE)
}
\keyword{effect}
\keyword{size}
