% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ncpt.R
\name{Ncpt}
\alias{Ncpt}
\title{Ncpt}
\usage{
Ncpt(x, q, df, confirm = FALSE)
}
\arguments{
\item{x}{sample mean}

\item{q}{population mean}

\item{df}{degrees of freedom}

\item{confirm}{defaults to FALSE}
}
\description{
Calculates non-central t
}
\examples{
Ncpt(x, q, df, confirm=FALSE)
}
\keyword{effect}
\keyword{size}
