% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.gen.SS.rm.R
\name{omega.gen.SS.rm}
\alias{omega.gen.SS.rm}
\title{Generalized Omega Squared for Multi-Way and Mixed ANOVA from F}
\usage{
omega.gen.SS.rm(dfm, dfe, ssm, ssm2, sst, mss, j, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{ssm}{sum of squares for the MAIN model/IV/between}

\item{ssm2}{sum sum of squares for the OTHER model/IV/between}

\item{sst}{sum of squares total across the whole ANOVA}

\item{mss}{mean square for the subject variance}

\item{j}{number of levels in the OTHER IV}

\item{Fvalue}{F statistic from the output for your IV}

\item{a}{significance level}
}
\value{
Provides omega squared with associated confidence intervals
and relevant statistics.

\item{omega}{omega squared}
\item{omegalow}{lower level confidence interval of omega}
\item{omegahigh}{upper level confidence interval of omega}
\item{dfm}{degrees of freedom for the model/IV/between}
\item{dfe}{degrees of freedom for the error/resisual/within}
\item{F}{F-statistic}
\item{p}{p-value}
}
\description{
This function displays generalized omega squared from ANOVA analyses
and its non-central confidence interval based on the F distribution.
This formula is appropriate for multi-way repeated measures designs and mix level designs.
}
\examples{
omega.gen.SS.rm(dfm = 2, dfe = 100,
                     ssm = 5339, ssm2 = 432, sst = 10947,
                     mss = 543, j = 3, Fvalue = 12, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{omega,}
\keyword{size,}
