% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.partial.SS.bn.R
\name{omega.partial.SS.bn}
\alias{omega.partial.SS.bn}
\title{omega.partial.SS.bn}
\usage{
omega.partial.SS.bn(dfm, dfe, msm, mse, ssm, n, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{msm}{mean square for the model/IV/between}

\item{mse}{mean square for the error/residual/within}

\item{ssm}{sum of squares for the model/IV/between}

\item{n}{total sample size}

\item{a}{significance level}
}
\description{
This function displays omega squared from ANOVA analyses
and its non-central confidence interval based on the F distribution.
This formula is appropriate for multi-way between subjects designs.
}
\examples{
omega.partial.SS.bn(dfm = 2, dfe = 100, msm = 214, mse = 100, ssm = 5339, n = 150, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{omega,}
\keyword{size,}
