% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.correl.R
\name{r.correl}
\alias{r.correl}
\title{r.correl}
\usage{
r.correl(r, n, a = 0.05, k = 2)
}
\arguments{
\item{r}{correlation coefficient}

\item{n}{sample size}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays transformation from r to r2
}
\examples{
r.correl(r = .5, n = 100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
