% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm1-data.R
\docType{data}
\name{rm1_data}
\alias{rm1_data}
\title{Repeated Measures Oneway ANOVA Example Data}
\format{A data frame including ratings toward pictures.

neutral: pulse during exposure to neutral stimuli
positive: pulse during exposure to positive stimuli
negative: pulse during exposure to negative stimuli}
\usage{
data(rm1_data)
}
\description{
Dataset for use in \code{\link{omega.F}}.
Participants were tested over several days to measure
variations in their pulse given different types of stimuli.
One stimulus was a neutral picture (like a toaster),
while other stimuli were cute/happy pictures (puppies, babies),
and negative stimuli (mutilated faces, pictures of war).
Were there differences in pulse for each participant
across the stimuli?
}
\keyword{datasets}
