% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm2-data.R
\docType{data}
\name{rm2_data}
\alias{rm2_data}
\title{Repeated Measures Two-way ANOVA Example Data}
\format{A data frame of ratings of word pair relation

group: A between-subjects variable indicating the type of instructions
subject: A subject number
fsglobsglo: A repeated measures condition of low FSG-BSG
fsghihbsglo: A repeated measures condition of high FSG, low BSG
fsglobsghi: A repeated measures condition of low FSG, high BSG
fsghibsghi: A repeated measures condition of high FSG-BSG}
\usage{
data(rm2_data)
}
\description{
Dataset for use in \code{\link{omega.partial.SS.rm}} and other
repeated measures ANOVA designs. This dataset includes a group variable
used for mixed repeated measures designs, a subject number, and two
repeated measures variables. These variables include FSG (forward strength)
which is a measure of the relation between two words like cheddar to cheese.
The second variable is BSG (backward strength), which is the opposite
relation (cheese to cheddar). Participants rated those word pairs in and the
strength of FSG and BSG was manipulated to measure overestimation of strength.
}
\keyword{datasets}
