% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singt-data.R
\docType{data}
\name{singt_data}
\alias{singt_data}
\title{One-Sample t-Test Example Data}
\format{
A data frame with 1 variable:
\describe{
  \item{SATscore}{Numeric. SAT scores of gifted/honors program students at one school.}
}
}
\usage{
data(singt_data)
}
\description{
Simulated dataset of SAT scores from gifted/honors students at a specific school,
intended for comparison to the national average SAT score (1080) for gifted/honors
students nationwide. Designed for use with functions such as \code{\link{d.single.t}}
and \code{\link{d.single.t.t}}.
}
\keyword{datasets}
