% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.chi.sq.R
\name{v.chi.sq}
\alias{v.chi.sq}
\title{v.chi.sq}
\usage{
v.chi.sq(x2, n, r, c, a = 0.05, k = 2)
}
\arguments{
\item{x2}{chi-square metric}

\item{n}{sample size}

\item{r}{correlation}

\item{c}{number of columns in the contigency table}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays the effect size and confidence interval for the specified chi-square metric using Cramer's V.
}
\examples{
v.chi.sq(x2, n, r, c, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
