% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa}
\alias{apa}
\title{apa}
\usage{
apa(value, decimals, leading)
}
\arguments{
\item{value}{input value to convert}

\item{decimals}{input how many decimal places}
  
\item{leading}{input T for leading decimals F for no leading decimals}
}
\description{
This function formats for APA style
}
\examples{
apa(value, decimals, leading)
}
\keyword{APA}
