% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa}
\alias{apa}
\title{apa}
\usage{
apa(value, decimals = 3, leading = T)
}
\arguments{
\item{value}{the number you want to format}

\item{decimals}{the number of decimal points desired}

\item{leading}{logical TRUE for leading zeroes on decimals and FALSE for no leading zeroes on decimals}
}
\description{
A function that formats decimals and leading zeroes
}
\examples{
Apa(x,k)
}
\keyword{APA,}
\keyword{decimals,}
\keyword{formatting}
