% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn1-data.R
\docType{data}
\name{bn1_data}
\alias{bn1_data}
\title{Between Subjects One-way ANOVA Example Data}
\format{A data frame of ratings of close interpersonal attachments

poor: individuals in poor health
fair: individuals in fair health
excellent: individuals in excellent health}
\usage{
data(bn1_data)
}
\description{
Dataset for use in \code{\link{eta.F}}, \code{\link{eta.full.SS}},
\code{\link{omega.F}}, \code{\link{omega.full.SS}},
and \code{\link{epsilon.full.SS}}, including ratings of
inter-personal attachments of 45-year-olds categorized as
being in excellent, fair, or poor health.
}
\references{
Nolan and Heizen Statistics for the Behavioral Sciences
(\href{https://www.macmillanlearning.com/Catalog/product/statisticsforthebehavioralsciences-rentalonly-fourthedition-nolan}{Book Link})
}
\keyword{datasets}
