% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn2-data.R
\docType{data}
\name{bn2_data}
\alias{bn2_data}
\title{Between Subjects Two-way ANOVA Example Data}
\format{A data frame of ratings of close interpersonal attachments

coach: an old or new coach
type: varying sports - baseball, basketball, football, soccer, volleyball
money: athletic spending (in thousands of dollars)}
\usage{
data(bn2_data)
}
\description{
Dataset for use in \code{\link{omega.partial.SS.bn}},
\code{\link{eta.partial.SS}}, and other between-subject's ANOVA
designs. This data includes (fake) atheletic budgets for baseball,
basketball, football, soccer, and volleyball teams
with new and old coaches to determine if there are differences in
spending across coaches and sports.
}
\keyword{datasets}
