% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq-data.R
\docType{data}
\name{chisq_data}
\alias{chisq_data}
\title{Chi-Square Example Data}
\format{A data frame of number of friends and number of children

friends: number of reported friends
kids: number of children}
\usage{
data(chisq_data)
}
\description{
Dataset for use in \code{\link{v.chi.sq}},
Individuals were polled and asked to report their number of friends
(low, medium, high) and number of kids (1, 2, 3+)
to determine if there was a relationship between friend
groups and number of children. It was hypothesized that
those with more children may have less time for
friendship maintaining activities.
}
\references{
Nolan and Heizen Statistics for the Behavioral Sciences
(\href{https://www.macmillanlearning.com/Catalog/product/statisticsforthebehavioralsciences-rentalonly-fourthedition-nolan}{Book Link})
}
\keyword{datasets}
