% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.avg.R
\name{d.dep.t.avg}
\alias{d.dep.t.avg}
\title{d for Dependent t with Average SD Denominator}
\usage{
d.dep.t.avg(m1, m2, sd1, sd2, n, a = 0.05)
}
\arguments{
\item{m1}{mean from first level}

\item{m2}{mean from second level}

\item{sd1}{standard deviation from first level}

\item{sd2}{standard deviation from second level}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
average standard deviation of each level as the denominator.
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
