% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.avg.R
\name{d.dep.t.avg}
\alias{d.dep.t.avg}
\title{d for Dependent t with Average SD Denominator}
\usage{
d.dep.t.avg(m1, m2, sd1, sd2, n, a = 0.05)
}
\arguments{
\item{m1}{mean from first level}

\item{m2}{mean from second level}

\item{sd1}{standard deviation from first level}

\item{sd2}{standard deviation from second level}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
average standard deviation of each level as the denominator.
}
\details{
Provided information is:
\itemize{
  \item d: the effect size
  \item dlow: the lower level confidence interval d value
  \item dhigh: the upper level confidence interval d value
  \item M1/M2: mean one and two
  \item M1low/M2low: the lower level confidence interval of mean one or two
  \item M1high/M2high: the upper level confidence interval of mean one or two
  \item sd1/sd2: the standard deviation of mean one and two
  \item se1/se2: the standard error of mean one and two
  \item n: the sample size
  \item df: the degrees of freedom (sample size - 1)
  }
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
