% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.avg.R
\name{d.dep.t.avg}
\alias{d.dep.t.avg}
\title{d for Dependent t with Average SD Denominator}
\usage{
d.dep.t.avg(m1, m2, sd1, sd2, n, a = 0.05)
}
\arguments{
\item{m1}{mean from first level}

\item{m2}{mean from second level}

\item{sd1}{standard deviation from first level}

\item{sd2}{standard deviation from second level}

\item{n}{sample size}

\item{a}{significance level}
}
\value{
The effect size (Cohen's d) with associated confidence intervals,
the confidence intervals associated with the means of each group,
standard deviations of the means for each group.

\item{d}{effect size}
\item{dlow}{the lower level confidence interval d value}
\item{dhigh}{the upper level confidence interval d value}
\item{M1/M2}{mean one and two}
\item{M1low/M2low}{the lower level confidence interval of mean one or two}
\item{M1high/M2high}{the upper level confidence interval of mean one or two}
\item{sd1/sd2}{the standard deviation of mean one and two}
\item{se1/se2}{the standard error of mean one and two}
\item{n}{the sample size}
\item{df}{the degrees of freedom (sample size - 1)}
}
\description{
This function displays d and the non-central confidence interval
for repeated measures data, using the average standard deviation of
each level as the denominator.
}
\details{
To calculate d, mean two is subtracted from mean one, which is
divided by the average standard deviation.
\href{Learn more on our example page.}{https://www.aggieerin.com/shiny-server/tests/deptavgm.html}
}
\examples{

#The following example is derived from the "dept_data" dataset included
in the MOTE library.

#In a study to test the effects of science fiction movies on people's
belief in the supernatural, seven people completed a measure of belief
in the supernatural before and after watching a popular science fiction
movie. Higher scores indicated higher levels of belief.

    t.test(dept_data$before, dept_data$after, paired = TRUE)

#You can type in the numbers directly, or refer to the dataset,
as shown below.

    d.dep.t.avg(m1 = 5.571, m2 = 4.429, sd1 = 1.988,
                sd2 = 2.878, n = 14, a = .05)

    d.dep.t.avg(5.571, 4.429, 1.988, 2.878, 14, .05)

    d.dep.t.avg(mean(dept_data$before), mean(dept_data$after),
                sd(dept_data$before), sd(dept_data$after),
                length(dept_data$before), .05)

#The mean measure of belief on the pretest (dept_data$before)
was 5.57, with a standard deviation of 1.99. The posttest
(dept_data$after) scores appeared lower (M = 4.43, SD = 2.88) but did not reach
significance, (t(7) = 1.1429, p = .20, d = .54), likely due to the small sample size
The effect size was moderate (d = 0.47), suggesting the movie may
have influenced belief in the supernatural.
}
\keyword{average,}
\keyword{cohen's}
\keyword{d}
\keyword{d,}
\keyword{dependent}
\keyword{effect}
\keyword{measures}
\keyword{paired-sample,}
\keyword{repeated}
\keyword{size,}
\keyword{t-test,}
