% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.diff.R
\name{d.dep.t.diff}
\alias{d.dep.t.diff}
\title{d for Dependent t with SD Difference Scores Denominator}
\usage{
d.dep.t.diff(mdiff, sddiff, n, a = 0.05)
}
\arguments{
\item{mdiff}{mean difference score}

\item{sddiff}{standard deviation of the difference scores}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
standard deviation of the differences as the denominator.
}
\details{
formula here
}
\examples{
d.dep.t.diff(mdiff = 5, sddiff = 3, n = 100, a = .05)
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
