% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.diff.R
\name{d.dep.t.diff}
\alias{d.dep.t.diff}
\title{d for Dependent t with SD Difference Scores Denominator}
\usage{
d.dep.t.diff(mdiff, sddiff, n, a = 0.05)
}
\arguments{
\item{mdiff}{mean difference score}

\item{sddiff}{standard deviation of the difference scores}

\item{n}{sample size}

\item{a}{significance level}
}
\value{
The effect size and associated confidence intervals, mean
differences with associated confidence intervals, standard
deviation of the differences, standard error, sample size, degrees of
freedom, the t-statistic, and the p-value.
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
standard deviation of the differences as the denominator.
}
\examples{
#In a study to test the effects of science fiction movies on people's
belief in the supernatural, seven people completed a measure of belief
in the supernatural before and after watching a popular science fiction movie.
Higher scores indicated high levels of belief. The mean difference score was 1.143,
while the standard deviation of the difference scores was 2.116 (the standard error difference was .800).
You can type in the numbers directly:
d.dep.t.diff(mdiff = 1.143, sddiff = 2.116, n = 14, a = .05)
or you can calculate from the data.

}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
