% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.diff.R
\name{d.dep.t.diff}
\alias{d.dep.t.diff}
\title{d for Dependent t with SD Difference Scores Denominator}
\usage{
d.dep.t.diff(mdiff, sddiff, n, a = 0.05)
}
\arguments{
\item{mdiff}{mean difference score}

\item{sddiff}{standard deviation of the difference scores}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
standard deviation of the differences as the denominator.
}
\details{
Provided information is:

\itemize{
  \item d: the effect size
  \item dlow: the lower level confidence interval d value
  \item dhigh: the upper level confidence interval d value
  \item mdiff: mean of the differences
  \item Mlow: the lower level confidence interval of mean
  \item Mhigh: the upper level confidence interval of mean
  \item sddiff: standard deviation of the differences
  \item se: the standard error
  \item n: the sample size
  \item df: the degrees of freedom (sample size - 1)
  \item t: t-statistic
  \item p: p-value
  }
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
