% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.diff.t.R
\name{d.dep.t.diff.t}
\alias{d.dep.t.diff.t}
\title{d from t for Repeated Measures with SD Difference Scores Denominator}
\usage{
d.dep.t.diff.t(t, n, a = 0.05)
}
\arguments{
\item{t}{t-test value}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data
and the non-central confidence interval using the
standard deviation of the differences as the denominator
estimating from the t-statistic.
}
\examples{
d.dep.t.diff.t(t = 5, n = 100, a = .05)
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
