% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.dep.t.rm.R
\name{d.dep.t.rm}
\alias{d.dep.t.rm}
\title{d.dep.t.rm}
\usage{
d.dep.t.rm(m1, m2, sd1, sd2, r, n, a = 0.05)
}
\arguments{
\item{m1}{mean from first level}

\item{m2}{mean from second level}

\item{sd1}{standard deviation from first level}

\item{sd2}{standard deviation from second level}

\item{r}{correlation between first and second level}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d for repeated measures data controlling for r
and the non-central confidence interval using the
average standard deviation of each level as the denominator.
}
\examples{
d.dep.t.rm(m1 = 20, m2 = 17, sd1 = 4, sd2 = 5, r = .30, n = 100, a = .05)
}
\keyword{dependent}
\keyword{effect}
\keyword{size,}
\keyword{t-test}
