% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.ind.t.t.R
\name{d.ind.t.t}
\alias{d.ind.t.t}
\title{d from t for Between Subjects}
\usage{
d.ind.t.t(t, n1, n2, a = 0.05)
}
\arguments{
\item{t}{t-test value}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\value{
Provides the effect size (Cohen's d) with associated confidence intervals,
degrees of freedom, t-statistic, and p-value.

\item{d}{effect size}
\item{dlow}{lower level confidence interval of d value}
\item{dhigh}{upper level confidence interval of d value}
\item{n1}{sample size}
\item{n2}{sample size}
\item{df}{degrees of freedom (n1 - 1 + n2 - 1)}
\item{t}{t-statistic}
\item{p}{p-value}
}
\description{
This function displays d for between subjects data
and the non-central confidence interval estimating from the t-statistic.
}
\details{
To calculate the t-statistic, mean two is subtracted from mean one and divided
by the standard error.

To calculate d, the t-statistic is divided by the square root of the degrees of freedom
and multiplied by two.

\href{Learn more on our example page.}{https://www.aggieerin.com/shiny-server/tests/indtt.html}
}
\examples{

#The following example is derived from the "indt_data" dataset, included
in the MOTE library.

#A forensic psychologist conducted a study to examine whether
being hypnotized during recall affects how well a witness
can remember facts about an event. Eight participants
watched a short film of a mock robbery, after which
each participant was questioned about what he or she had
seen. The four participants in the experimental group
were questioned while they were hypnotized. The four
participants in the control group recieved the same
questioning without hypnosis.

    hyp = t.test(correctq ~ group, data = indt_data)

#You can type in the numbers directly, or refer to the dataset,
as shown below.

    d.ind.t.t(t = -2.6599, n1 = 4, n2 = 4, a = .05)

    d.ind.t.t(-2.6599, 4, 4, .05)

    d.ind.t.t(hyp$statistic,
              length(indt_data[1:4, ('correctq')]),
              length(indt_data[5:8, ('correctq')]),
              .05)

#Contrary to the hypothesized result, the group that underwent hypnosis were
significantly less accurate while reporting facts than the control group
with a large effect size, t(7) = -2.66, p = .043, d = 2.17.

d.ind.t.t(t = 4.12, n1 = 100, n2 = 100, a = .05)
}
\keyword{effect}
\keyword{independent}
\keyword{size,}
\keyword{t}
