% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.prop.R
\name{d.prop}
\alias{d.prop}
\title{d.prop}
\usage{
d.prop(p1, p2, n1, n2, a = 0.05, k = 2)
}
\arguments{
\item{p1}{proportion of people from group 1}

\item{p2}{proportion of people from group 2}

\item{n1}{sample size group 1}

\item{n2}{sample size group 2}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays proportion differences, independent
}
\examples{
d.prop(p1 = .4, p2 = .6, n1 = 100, n2 = 100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
