% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d.single.t.R
\name{d.single.t}
\alias{d.single.t}
\title{d for Single t from Means}
\usage{
d.single.t(m, u, sd, n, a = 0.05)
}
\arguments{
\item{m}{sample mean}

\item{u}{population mean}

\item{sd}{sample standard deviation}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays d and non-central confidence interval for single t from means.
}
\examples{
d.single.t(m = 20, u = 17, sd = 4, n = 100, a = .05)
}
\keyword{effect}
\keyword{single}
\keyword{size,}
\keyword{t}
