% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Delta.ind.t.R
\name{Delta.ind.t}
\alias{Delta.ind.t}
\title{Delta.ind.t}
\usage{
Delta.ind.t(m1, m2, sd1, sd2, n1, n2, a = 0.05, k = 2)
}
\arguments{
\item{m1}{mean from control group}

\item{m2}{mean from experimental group}

\item{sd1}{standard deviation from control group}

\item{sd2}{standard deviation from experimental group}

\item{n1}{sample size from control group}

\item{n2}{sample size from experimental group}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function displays cohen's d and confidence interval for dependent t from averages
}
\examples{
Delta.ind.t(m1=20, m2=17, sd1=4, sd2=5, n1=100, n2 = 100, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
