% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.anova.test.R
\name{eta.anova}
\alias{eta.anova}
\title{eta.anova}
\usage{
eta.anova(dfm, dfe, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom model}

\item{dfe}{degrees of freedom error}

\item{Fvalue}{F statistic}

\item{a}{significance level}
}
\description{
This function displays eta, r squared, ICCs from ANOVA analyses
and their non-central confidence interval based on the F distribution.
}
\examples{
eta.anova(dfm = 2, dfe = 20, Fvalue = 5.7, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{eta,}
\keyword{size,}
