% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta.F.R
\name{eta.F}
\alias{eta.F}
\title{eta.F}
\usage{
eta.F(dfm, dfe, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{Fvalue}{F statistic}

\item{a}{significance level}
}
\description{
This function displays eta, r squared, ICCs from ANOVA analyses
and their non-central confidence interval based on the F distribution.
These values are calculated directly from F statistics and can be used
for between subjects and repeated measures designs.
Remember if you have two or more IVs, these values are partial eta squared.
}
\examples{
eta.F(dfm = 2, dfe = 20, Fvalue = 5.7, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{eta,}
\keyword{size,}
