% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.ind.t.R
\name{g.ind.t}
\alias{g.ind.t}
\title{d-g Corrected for Independent t}
\usage{
g.ind.t(m1, m2, sd1, sd2, n1, n2, a = 0.05)
}
\arguments{
\item{m1}{mean group one}

\item{m2}{mean group two}

\item{sd1}{standard deviation group one}

\item{sd2}{standard deviation group two}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\value{
D-g corrected with associated confidence intervals,
the confidence intervals associated with the means of each group,
standard deviations of the means for each group, relevant statistics.

\item{d}{d-g corrected effect size}
\item{dlow}{lower level confidence interval d-g corrected}
\item{dhigh}{upper level confidence interval d-g corrected}
\item{M1}{mean group one}
\item{sd1}{standard deviation of group one}
\item{se1}{standard error of group one}
\item{M1low}{lower level confidence interval of mean one}
\item{M1high}{upper level confidence interval of mean one}
\item{M2}{mean two}
\item{sd2}{standard deviation of mean two}
\item{se1}{standard error of mean two}
\item{M2low}{lower level confidence interval of mean two}
\item{M2high}{upper level confidence interval of mean two}
\item{spooled}{pooled standard deviation}
\item{sepooled}{pooled standard error}
\item{correction}{g corrected}
\item{n1}{size of sample one}
\item{n2}{size of sample two}
\item{df}{degrees of freedom}
\item{t}{t-statistic}
\item{p}{p-value}
}
\description{
This function displays d-g corrected
and the non-central confidence interval for independent t.
}
\details{
The correction is calculated by dividing three by the sum of both
sample sizes after multiplying by four and subtracting nine.
This amount is deducted from one.

correction = 1 - (3 / (4 * (n1 + n2) - 9))

D-g corrected is calculated by substracting mean two from mean one,
dividing by the pooled standard deviation which is multiplied
by the correction above.

d-g corrected = ((m1 - m2) / spooled) * correction

\href{https://www.aggieerin.com/shiny-server/tests/indtg.html}{Learn more on our example page.}
}
\examples{
g.ind.t(m1 = 20, m2 = 17, sd1 = 4, sd2 = 5, n1 = 100, n2 = 100, a = .05)
}
\keyword{correction}
\keyword{effect}
\keyword{independent}
\keyword{size,}
\keyword{t,}
