% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.ind.t.R
\name{g.ind.t}
\alias{g.ind.t}
\title{g.ind.t}
\usage{
g.ind.t(m1, m2, sd1, sd2, n1, n2, a = 0.05)
}
\arguments{
\item{m1}{mean group one}

\item{m2}{mean group two}

\item{sd1}{standard deviation group one}

\item{sd2}{standard deviation group two}

\item{n1}{sample size group one}

\item{n2}{sample size group two}

\item{a}{significance level}
}
\description{
This function displays d-g corrected 
and the non-central confidence interval for independent t.
}
\examples{
g.ind.t(m1 = 20, m2 = 17, sd1 = 4, sd2 = 5, n1 = 100, n2 = 100, a = .05)
}
\keyword{correction}
\keyword{effect}
\keyword{independent}
\keyword{size,}
\keyword{t,}
