% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ges.partial.SS.rm.R
\name{ges.partial.SS.rm}
\alias{ges.partial.SS.rm}
\title{Partial Generalized Eta-Squared for ANOVA from F}
\usage{
ges.partial.SS.rm(dfm, dfe, ssm, sss, sse1, sse2, sse3, Fvalue, a = 0.05)
}
\arguments{
\item{dfm}{degrees of freedom for the model/IV/between}

\item{dfe}{degrees of freedom for the error/residual/within}

\item{ssm}{sum of squares for the model/IV/between}

\item{sss}{sum of squares subject variance}

\item{sse1}{sum of squares for the error/residual/within for the first IV}

\item{sse2}{sum of squares for the error/residual/within for the second IV}

\item{sse3}{sum of squares for the error/residual/within for the interaction}

\item{Fvalue}{F statistic}

\item{a}{significance level}
}
\value{
Partial generalized eta-squared (GES) with associated confidence intervals
and relevant statistics.
\item{ges}{effect size}
\item{geslow}{lower level confidence interval for ges}
\item{geshigh}{upper level confidence interval for ges}
\item{dfm}{degrees of freedom for the model/IV/between}
\item{dfe}{degrees of freedom for the error/residual/within}
\item{F}{F-statistic}
\item{p}{p-value}
}
\description{
This function displays partial ges squared from ANOVA analyses
and its non-central confidence interval based on the F distribution.
This formula works for multi-way repeated measures designs.
}
\details{
To calculate partial generalized eta squared, first, the sum of
squares of the model, sum of squares of the subject
variance, sum of squares for the first and second independent variables,
and the sum of squares for the interaction are added together.
The sum of squares of the model is divided by this value.

ges <- ssm / (ssm + sss + sse1 + sse2 + sse3)

\href{https://www.aggieerin.com/shiny-server/tests/gesrmss.html}{Learn more on our example page.}
}
\examples{
ges.partial.SS.rm(dfm = 2, dfe = 100, ssm = 435, sss = 659, sse1 = 435, sse2 = 446, sse3 = 546, Fvalue = 5.46, a = .05)
}
\keyword{ANOVA}
\keyword{effect}
\keyword{ges,}
\keyword{size,}
