% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{odds}
\alias{odds}
\title{Odds Ratios}
\usage{
odds(n11, n12, n21, n22, a = 0.05)
}
\arguments{
\item{n11}{sample size for level 1.1}

\item{n12}{sample size for level 1.2}

\item{n21}{sample size for level 2.1}

\item{n22}{sample size for level 2.2}

\item{a}{significance level}
}
\value{
Provides odds ratios with associated confidence intervals
and relevant statistics.

\item{odds}{odds statistic}
\item{olow}{lower level confidence interval of odds statistic}
\item{ohigh}{upper level confidence interval of odds statistic}
\item{se}{standard error}
}
\description{
This function displays odds ratios and their normal confidence intervals.
}
\details{
This statistic is the ratio between level 1.1 divided by level 1.2, and
level 2.1 divided by 2.2. In other words, these are the odds of level 1.1
given level 1 overall versus level 2.1 given level 2 overall.
\href{https://www.aggieerin.com/shiny-server/tests/chio.html}{Learn more on our example page.}
}
\examples{
odds(n11 = 10, n12 = 15, n21 = 20, n22 = 5, a = .05)
}
\keyword{effect}
\keyword{odds}
\keyword{ratios}
\keyword{size,}
