% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Odds.R
\name{Odds}
\alias{Odds}
\title{Odds}
\usage{
Odds(n11, n12, n21, n22, a = 0.05, k = 2)
}
\arguments{
\item{n11}{input value}

\item{n12}{input value}

\item{n21}{input value}

\item{n22}{input value}

\item{a}{significance level}

\item{k}{significant digits to use for formatting}
}
\description{
This function Displays odds ratios, sensitivity / specificity
}
\examples{
Odds(n11, n12, n21, n22, a = .05, k = 2)
}
\keyword{effect}
\keyword{size}
