% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r.correl.R
\name{r.correl}
\alias{r.correl}
\title{r.correl}
\usage{
r.correl(r, n, a = 0.05)
}
\arguments{
\item{r}{correlation coefficient}

\item{n}{sample size}

\item{a}{significance level}
}
\description{
This function displays transformation from r to r2 to calculate
the non-central confidence interval for r2 using the F distribution.
}
\examples{
r.correl(r = .5, n = 100, a = .05)
}
\keyword{correlation}
\keyword{effect}
\keyword{size,}
