% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singt-data.R
\docType{data}
\name{singt_data}
\alias{singt_data}
\title{Single Sample t Example Data}
\format{A data frame including a single sample consisting of SAT scores of students from
a gifted/honors program at a specific school.}
\usage{
data(singt_data)
}
\description{
A simulated dataset for use in \code{\link{d.single.t}} and \code{\link{d.single.t.t}}, including
gifted/honors student SAT scores from a specific school to use for comparison with the
national average SAT score (1080) of gifted/honors students nationwide.
}
\keyword{datasets}
