% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v.chi.sq.R
\name{v.chi.sq}
\alias{v.chi.sq}
\title{v.chi.sq}
\usage{
v.chi.sq(x2, n, r, c, a = 0.05, k = 2)
}
\arguments{
\item{x2}{chi-square statistic}

\item{n}{sample size}

\item{r}{number of rows in the contingency table}

\item{c}{number of columns in the contingency table}

\item{a}{significance level}
}
\description{
This function displays V and non-central confidence interval 
for the specified chi-square statistic.
}
\examples{
v.chi.sq(x2 = 4, n = 25, r = 2, c = 2, a = .05)
}
\keyword{chi-square}
\keyword{effect}
\keyword{size,}
